Ejercicio 6

+++ app/src/main/java/cl/awakelab/ejercicioindividual5modulo6/data/remote/Terreno.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 class Terreno(
     val id: String,
-    @SerializedName("price") val precio: Int,
+    @SerializedName("price") val precio: Long,
     @SerializedName("type") val tipo: String,
     @SerializedName("img_src") val imagen: String)
